@@ -1,9 +1,9 @@
 module github.com/TelefonicaTC2Tech/golium
 
-go 1.15
+go 1.16
 
 require (
-	github.com/AdguardTeam/dnsproxy v0.37.0
+	github.com/AdguardTeam/dnsproxy v0.41.2
 	github.com/aws/aws-sdk-go v1.38.71
 	github.com/cucumber/godog v0.12.0
 	github.com/elastic/go-elasticsearch/v7 v7.12.0
@@ -11,7 +11,7 @@ require (
 	github.com/google/uuid v1.2.0
 	github.com/kelseyhightower/envconfig v1.4.0
 	github.com/lestrrat-go/jwx v1.1.4
-	github.com/miekg/dns v1.1.40
+	github.com/miekg/dns v1.1.44
 	github.com/pkg/errors v0.9.1
 	github.com/sirupsen/logrus v1.8.0
 	github.com/spf13/pflag v1.0.5
@@ -21,7 +21,5 @@ require (
 	github.com/tidwall/sjson v1.2.4
 	github.com/xeipuuv/gojsonpointer v0.0.0-20190905194746-02993c407bfb // indirect
 	github.com/xeipuuv/gojsonschema v1.2.0
-	golang.org/x/sys v0.0.0-20210119212857-b64e53b001e4 // indirect
-	golang.org/x/text v0.3.5 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b
 )
